requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
gunicorn>=21.0.0

# Database
sqlalchemy>=2.0.30
alembic>=1.13.2
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Validation
pydantic>=2.8.0
pydantic-settings>=2.4.0
email-validator>=2.2.0

# AI Integration (optional)
# google-generativeai>=0.7.0

# Utilities
python-dotenv>=1.0.0

# CORS
# Already included in FastAPI

